Actually remove the old params
